SQLi Timeout Patch
I noticed that the tool is riddled with SQLi false positives, and I've made a determination that the ResponseTime() function is the reason that it's creating a ton of false positives. Bumping it to 15 seconds worked about half of the time, but 20 seconds seemed like a fair change that still found true positives and reduced false positives significantly.